fixed roles for mod take 3

+++ app/controllers/topics_controller.rb
@@ -2,9 +2,8 @@ class TopicsController < ApplicationController
   # use the before action filter and require_sign_in method to redirect users who attempt to access actions other than index and show
   before_action :require_sign_in, except: [:index, :show]
   # use another before_action filter to check role of signed in users.
-  before_action :authorize_mod, only: [:edit], except: [:index, :show]
 
-  before_action :authorize_user, only: [:create, :new, :delete], except: [:index, :show]
+  # before_action :authorize_user, except: [:index, :show]
 
   def index
     @topics = Topic.all
@@ -16,10 +15,21 @@ class TopicsController < ApplicationController
 
   def new
     @topic = Topic.new
+
+    unless current_user.admin?
+      flash[:alert] = "You must be an admin to do that."
+      redirect_to topics_path
+    end
+
   end
 
   def create
     @topic = Topic.new(topic_params)
+
+    unless current_user.admin?
+      flash[:alert] = "You must be an admin to do that."
+      redirect_to topics_path
+    end
 
     if @topic.save
       redirect_to @topic, notice: "Topic was saved successfully"
@@ -30,12 +40,24 @@ class TopicsController < ApplicationController
 
   def edit
     @topic = Topic.find(params[:id])
+
+
+    unless current_user.moderator? || current_user.admin?
+      flash[:alert] = "You are not authorized to do that."
+      redirect_to topics_path
+    end
+
   end
 
   def update
     @topic = Topic.find(params[:id])
 
     @topic.assign_attributes(topic_params)
+
+    unless current_user.moderator? || current_user.admin?
+      flash[:alert] = "You are not authorized to do that."
+      redirect_to topics_path
+    end
 
     if @topic.save
       flash[:notice] = "Topic was updated."
@@ -48,6 +70,12 @@ class TopicsController < ApplicationController
 
   def destroy
     @topic = Topic.find(params[:id])
+
+    unless current_user.admin?
+      flash[:alert] = "You must be an admin to do that."
+      redirect_to topics_path
+    end
+
 
     if @topic.destroy
       flash[:notice] = "\"#{@topic.name}\" was deleted successfully"
@@ -63,18 +91,18 @@ class TopicsController < ApplicationController
   def topic_params
     params.require(:topic).permit(:name, :description, :public)
   end
+  #
+  # def authorize_mod
+  #   unless current_user.moderator? || current_user.admin?
+  #     flash[:alert] = "You must be a moderator or admin to do that."
+  #     redirect_to topics_path
+  #   end
+  # end
 
-  def authorize_mod
-    unless current_user.moderator? || current_user.admin?
-      flash[:alert] = "You must be a moderator or admin to do that."
-      redirect_to topics_path
-    end
-  end
-
-  def authorize_user
-    unless current_user.admin?
-      flash[:alert] = "You must be an admin to do that."
-      redirect_to topics_path
-    end
-  end
+  # def authorize_user
+  #   unless current_user.admin?
+  #     flash[:alert] = "You must be an admin to do that."
+  #     redirect_to topics_path
+  #   end
+  # end
 end
